# MonolithApi/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Monolith api",
    "description": "This is an api of our monolith backend for our Architecture class (M1, Expert Dev Logiciel Mob Iot, YNOV)",
    "contact": {
      "name": "Sada Demba Thiam",
      "url": "https://github.com/SadaDemba"
    },
    "version": "v1"
  },
  "paths": {
...
    "/api/ProductReductions/Shop/{shopId}": {
      "get": {
        "tags": [
          "ProductReductions"
        ],
        "parameters": [
          {
            "name": "shopId",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProductReduction"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProductReduction"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProductReduction"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/ProductReductions/Paginated/Shop/{shopId}": {
      "get": {
        "tags": [
          "ProductReductions"
        ],
        "parameters": [
          {
            "name": "shopId",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "pageNumber",
            "in": "query",
            "style": "form",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "style": "form",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProductReductionResponseResource"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductReductionResponseResource"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductReductionResponseResource"
                }
              }
            }
          }
        }
      }
    },
    "/api/ProductReductions/Paginated": {
      "get": {
        "tags": [
          "ProductReductions"
        ],
        "parameters": [
          {
            "name": "pageNumber",
            "in": "query",
            "style": "form",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "style": "form",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProductReductionResponseResource"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductReductionResponseResource"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductReductionResponseResource"
                }
              }
            }
          }
        }
      }
    },
    "/api/ProductReductions/Shop/{shopId}/Product/{productId}": {
      "get": {
        "tags": [
          "ProductReductions"
        ],
        "parameters": [
          {
            "name": "shopId",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "productId",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "pageNumber",
            "in": "query",
            "style": "form",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "style": "form",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProductReductionResponseResource"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductReductionResponseResource"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductReductionResponseResource"
                }
              }
            }
          }
        }
      }
    },
    "/api/ProductReductions/Product/{productId}": {
      "get": {
        "tags": [
          "ProductReductions"
        ],
        "parameters": [
          {
            "name": "productId",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "pageNumber",
            "in": "query",
            "style": "form",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "style": "form",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProductReductionResponseResource"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductReductionResponseResource"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductReductionResponseResource"
                }
              }
            }
          }
        }
      }
    },
    "/api/ProductReductions/IsActivated": {
      "get": {
        "tags": [
          "ProductReductions"
        ],
        "parameters": [
          {
            "name": "isActivated",
            "in": "query",
            "style": "form",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "pageNumber",
            "in": "query",
            "style": "form",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "style": "form",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProductReductionResponseResource"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductReductionResponseResource"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductReductionResponseResource"
                }
              }
            }
          }
        }
      }
    },
    "/api/ProductReductions/Shop/{shopId}/IsActivated": {
      "get": {
        "tags": [
          "ProductReductions"
        ],
        "parameters": [
          {
            "name": "shopId",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "isActivated",
            "in": "query",
            "style": "form",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "pageNumber",
            "in": "query",
            "style": "form",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "style": "form",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProductReductionResponseResource"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductReductionResponseResource"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductReductionResponseResource"
                }
              }
            }
          }
        }
      }
    },
    "/api/ProductReductions/Reduction/{id}": {
      "get": {
        "tags": [
          "ProductReductions"
        ],
        "parameters": [
          {
            "name": "reductionId",
            "in": "query",
            "style": "form",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "shopId",
            "in": "query",
            "style": "form",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "pageNumber",
            "in": "query",
            "style": "form",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "style": "form",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProductReduction"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductReduction"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductReduction"
                }
              }
            }
          }
        }
      }
    },
    "/api/ProductReductions/{id}": {
      "get": {
        "tags": [
          "ProductReductions"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProductReduction"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductReduction"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductReduction"
                }
              }
            }
          }
        }
      }
    },
    "/api/ProductReductions/Seller/{userId}/SetIsActivated/{id}": {
      "put": {
        "tags": [
          "ProductReductions"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "isActivated",
            "in": "query",
            "style": "form",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/ProductReductions/Seller/{userId}/{id}": {
      "put": {
        "tags": [
          "ProductReductions"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProductReduction"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ProductReduction"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ProductReduction"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "ProductReductions"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/ProductReductions/Seller/{userId}": {
      "post": {
        "tags": [
          "ProductReductions"
        ],
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProductReduction"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ProductReduction"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ProductReduction"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProductReduction"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductReduction"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductReduction"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/Products/Paginated": {
      "get": {
        "tags": [
          "Products"
        ],
        "parameters": [
          {
            "name": "pageNumber",
            "in": "query",
            "style": "form",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "style": "form",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Product"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Product"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Product"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/Products/{id}": {
      "get": {
        "tags": [
          "Products"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Product"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Product"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Product"
                }
              }
            }
          }
        }
      }
    },
    "/api/Products/Shop/{shopId}": {
      "get": {
        "tags": [
          "Products"
        ],
        "parameters": [
          {
            "name": "shopId",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "pageNumber",
            "in": "query",
            "style": "form",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "style": "form",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Product"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Product"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Product"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/Products/Seller/{userId}/{id}": {
      "put": {
        "tags": [
          "Products"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Product"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Product"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Product"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "Products"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Products/Seller/{userId}": {
      "post": {
        "tags": [
          "Products"
        ],
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Product"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Product"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Product"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Product"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Product"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Product"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/Reductions/Seller/{userId}/{id}": {
      "put": {
        "tags": [
          "Reductions"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Reduction"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Reduction"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Reduction"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "Reductions"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Reductions/Seller/{userId}": {
      "post": {
        "tags": [
          "Reductions"
        ],
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Reduction"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Reduction"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Reduction"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Reduction"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Reduction"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Reduction"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ProductReduction": {
        "required": [
          "isActivated",
          "productId",
          "reductionId"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "productId": {
            "type": "integer",
            "format": "int32"
          },
          "reductionId": {
            "type": "integer",
            "format": "int32"
          },
          "isActivated": {
            "type": "boolean"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "reduction": {
            "$ref": "#/components/schemas/Reduction"
          }
        },
        "additionalProperties": false
      },
      "ProductReductionResponseResource": {
        "type": "object",
        "properties": {
          "pageNumber": {
            "type": "integer",
            "format": "int32"
          },
          "pageSize": {
            "type": "integer",
            "format": "int32"
          },
          "firstPage": {
            "type": "string",
            "nullable": true
          },
          "lastPage": {
            "type": "string",
            "nullable": true
          },
          "totalPages": {
            "type": "integer",
            "format": "int32"
          },
          "totalRecords": {
            "type": "integer",
            "format": "int32"
          },
          "nextPage": {
            "type": "string",
            "nullable": true
          },
          "previousPage": {
            "type": "string",
            "nullable": true
          },
          "currentPage": {
            "type": "string",
            "nullable": true
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProductReduction"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
    }
  }
}